Fixed tests

+++ conftest.py
@@ -1,6 +1,5 @@
 from faker import Faker
 
-from data import CourierData
 from api.courier_api import CourierApi
 from helper import generate_random_string, generate_random_date
 
@@ -19,21 +18,19 @@ def courier_data():
     }
     
     yield courier_data
-    
-    create_response = CourierApi.create_courier(courier_data)
-    login_data = CourierData.get_login_password(courier_data, create_response)
-    login_response = CourierApi.login_courier(login_data)
-    
-    if login_response.status_code != 200:
-        print(f"Логин курьера не удался: {login_response.status_code}")
+    print(courier_data)
 
-    courier_id = CourierData.get_courier_id(login_response)
+@pytest.fixture(autouse=True)
+def cleanup_courier_after_test(request):
     
-    delete_response = CourierApi.delete_courier(courier_id)
+    def fin():
+        if hasattr(request.instance, 'created_couriers'):
+            for courier_id in request.instance.created_couriers:
+                CourierApi.delete_courier(courier_id)
+                print(f"Курьер ID={courier_id} удален")
     
-    if delete_response.status_code != 200:
-        print(f"Не удалось удалить курьера id={courier_id}")
-    
+    request.addfinalizer(fin)
+    yield
 
 @pytest.fixture
 def order_data():
@@ -51,3 +48,4 @@ def order_data():
     "color": ["BLACK"]
 }
     return order_data
+
